Add image and recipe details to beef rice page
Created a directory to store images named img within the main
project directory. Downloaded an image of the finished beef
rice dish and saved it to the img directory.
Relative link to the image added to the beef rice page.
Set image height and width with HTML.
Add headings and a list of ingredients to beef rice page.